requirements.txt
pytest==8.0.2
selene==2.0.0rc9
Appium-Python-Client==4.0.0
allure-pytest==2.13.5
python-dotenv==1.0.1
requests==2.31.0